requirements.txt
asgiref==3.10.0
coverage==7.11.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.1.0
packaging==25.0
pillow==12.0.0
pluggy==1.6.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-cov==7.0.0
pytest-django==4.11.1
pytest-pythonpath==0.7.3
sqlparse==0.5.3
